Lots of changes and refactor in this large commit (#67)
- Simplify a lot of the logic surrounding catalog, database and
settings.
- Refactor catalogs entirely to be SQLite files instead of JSON.
- Rewrote the schemas into their own NPM packages
- Rewrote schemas to use Zod instead of JSON Schema and AJV. This
simplifies everything, including building types (since Zod typing is
excellent).
- Added prettier rules and executed them, reformatting the entire
TypeScript.
- Added new functions and capabilities to the React frontend for 1FPGA.
- Added more tests (though still not enough coverage).
- Simplified a lot of the JavaScript, reducing the bundle by a bit less
than half (250kb before to ~140kb now).
- Rebuilt the catalog logic to be simpler; instead of having tables for
each part of catalogs, the catalog JSON is now kept directly in the
database. This simplifies logic and improves performance at the cost of
memory. Large catalogs might be a problem we will have to solve later.
- Moved from class-based architecture to more functional with named
imports which Rollup optimizes very well.

And a lot more.

UI and UX should have remained the same (with some minor advantages such
as a faster start time but a display of "initializing").

## BREAKING CHANGE
Unfortunately, it was much simpler to just nuke the database and rebuild
it from scratch, since a lot of the database schema was completely
refactored. This means people will lose their local game and settings
data.

This should not happen again, as the game identification and system
information is really the part that moves a lot, and will be moved to
external databases now.